Move StrippedDependency to file
